check return types

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -5,7 +5,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: int = 10) -> int:
+async def wait_random(max_delay: int = 10) -> float:
     """Waits for a random delay btwn 0 and max_delay and returns it"""
     num = random.uniform(0, max_delay)
     return num
